Restructured package to match the current internal standards for a worker package. Adapted installation instructions.
