+++ setup.cfg
@@ -3,16 +3,14 @@
 # http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
 
 [metadata]
-name = raster_binary_validation
-description = Add a short description here!
-author = florian.roth
-author-email = florian.roth@tuwien.ac.at
+name = raster-binary-validation
+description = Validation of binary classification results.
+author = TU Wien GEO MRS group
+author-email = remote.sensing@geo.tuwien.ac.at
 license = mit
 long-description = file: README.rst
 long-description-content-type = text/x-rst; charset=UTF-8
-url = https://github.com/pyscaffold/pyscaffold/
-project-urls =
-    Documentation = https://pyscaffold.org/
+url = https://git.eodc.eu/jrc-gfm/tuwien/gfm-sigma-offline
 # Change if running only on Windows, Mac or Linux (comma-separated)
 platforms = any
 # Add here all kinds of additional classifiers as defined under
@@ -30,7 +28,12 @@ package_dir =
 # DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
 setup_requires = pyscaffold>=3.3a0,<4
 # Add here dependencies of your project (semicolon/line-separated), e.g.
-# install_requires = numpy; scipy
+
+install_requires =
+    equi7grid
+    pandas
+    numpy
+
 # The usage of test_requires is discouraged, see `Dependency Management` docs
 # tests_require = pytest; pytest-cov
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
